+++ site/dat/docs/Changelog.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 1.6.4 (next)
+ - rely on Gatling simulation auto-detection mechanism when `simulation` field is not set
+ - fix 'non-existent scenario' error case
+ - fix config cleanup in cloud provisioning
+ - fix Selenium crash when used in multi-execution with a shared scenario
+
 ## 1.6.3 <sup>17 jun 2016</sup>
  - fix percentile value handling in passfail criteria
  - add setting up of CWD on server side